api/backboard-hello/pyproject.toml
[tool.poetry]
name = "backboard-hello"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.2.2"
backboard-sdk = "^1.5.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
